// client/dist/widgets/arcgis/query/src/config.ts
export enum CreateToolType {
  Point = 'Point',
  Polyline = 'Polyline',
  Polygon = 'Polygon',
  Rectangle = 'Rectangle',
  Circle = 'Circle'
}
// ...
export enum SpatialFilterType {
  CurrentMapExtent = 'CurrentMapExtent',
  InteractiveDrawMode = 'InteractiveDrawMode',
  None = ''
}

export enum SpatialRelation {
  Intersect = 'intersects',
  Contain = 'contains',
  Cross = 'crosses',
  EnvelopeIntersect = 'envelope-intersects',
  IndexIntersect = 'index-intersects',
  Overlap = 'overlaps',
  Touch = 'touches',
  Within = 'within',
}
// ...
export enum UnitType {
  Miles = 'Miles',
  Kilometers = 'Kilometers',
  Feet = 'Feet',
  Meters = 'Meters',
  NauticalMiles = 'NauticalMiles',
}

export const mapJSAPIUnitToDsUnit: { [key: string]: Units } = {
  [UnitType.Miles]: 'esriSRUnit_StatuteMile',
  [UnitType.Kilometers]: 'esriSRUnit_Kilometer',
  [UnitType.Feet]: 'esriSRUnit_Foot',
  [UnitType.Meters]: 'esriSRUnit_Meter',
  [UnitType.NauticalMiles]: 'esriSRUnit_NauticalMile'
}

export enum ListDirection {
  Horizontal = 'Horizontal',
  Vertical = 'Vertical',
}

export enum PagingType {
  MultiPage = 'MultiPage',
  LazyLoad = 'LazyLoad',
}

export enum FieldsType {
  PopupSetting = 'PopupSetting',
  SelectAttributes = 'SelectAttributes',
}

export enum SortDirection {
  Asc = 'Asc',
  Desc = 'Desc',
}

export enum ResultSelectMode {
  Single = 'Single',
  Multiple = 'Multiple',
}

export interface SortOption {
  name: string
  direction: SortDirection
}

export enum SymbolType {
  DefaultSymbol = 'DefaultSymbol',
  CustomSymbol = 'CustomSymbol',
}

// eslint-disable-next-line  @typescript-eslint/naming-convention
export interface QueryItemType {
  configId: string
  icon?: IconResult
  name?: string
  displayLabel?: boolean
  useDataSource?: UseDataSource
  outputDataSourceId?: string
  useAttributeFilter?: boolean
  useSpatialFilter?: boolean
  attributeFilterLabel?: string
  spatialFilterLabel?: string
  attributeFilterDesc?: string
  spatialFilterDesc?: string
  resultsLabel?: string
  sqlExprObj?: SqlExpression
  spatialMapWidgetIds?: string[]
  spatialFilterTypes?: SpatialFilterType[]
  spatialInteractiveCreateToolTypes?: CreateToolType[]
  spatialInteractiveEnableBuffer?: boolean
  spatialInteractiveBufferDistance?: number
  spatialInteractiveBufferUnit?: UnitType
  spatialRelations?: SpatialRelation[]
  spatialRelationUseDataSources?: UseDataSource[]
  spatialRelationEnableSelectTool?: boolean
  spatialRelationEnableBuffer?: boolean
  spatialRelationBufferDistance?: number
  spatialRelationBufferUnit?: UnitType
  spatialIncludeRuntimeData?: boolean
  resultListDirection?: ListDirection
  resultPagingStyle?: PagingType
  resultFieldsType?: FieldsType
  resultTitleFields?: string[]
  resultTitleExpression?: string
  resultDisplayFields?: string[]
  resultSymbolType?: SymbolType
  resultCustomSymbol?: any
  resultAllowChangeSymbol?: boolean
  resultExpandByDefault?: boolean
  resultSelectMode?: ResultSelectMode
  allowExport?: boolean
  sortOptions?: OrderByOption[]
  itemSizeMap?: {
    arrangementHorizontalPopper?: SizeMap
  }
}

export interface SizeMap {
  minSize?: Size
  defaultSize?: Size
}